chore: move properties to an internal package
We do not want to expose how the properties are read
